src/backend/requirements.txt
accelerate==0.33.0
annotated-types==0.7.0
anyio==4.6.0
autocorrect==2.6.1
benepar==0.2.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
distro==1.9.0
elastic-transport==8.13.1
elasticsearch==8.14.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
fsspec==2024.6.1
gensim==4.3.3
git-filter-repo==2.45.0
h11==0.14.0
hdbscan==0.8.38.post1
httpcore==1.0.6
httpx==0.25.2
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
inexactsearch==1.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
neo4j==5.25.0
networkx==3.2.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai==1.51.0
packaging==23.2
pandas==2.2.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.4.0
preshed==3.0.9
protobuf==5.27.3
psutil==6.0.0
pydantic==1.10.17
pydantic_core==2.20.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
silpa-common==0.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soundex==1.1.3
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
top2vec==1.0.34
torch==2.4.0
torch-struct==0.5
tqdm==4.66.5
transformers==4.44.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wordcloud==1.9.3
wrapt==1.16.0
zipp==3.19.2
neo4j==5.25.0
langchain==0.1.20
langchain-community==0.0.38
langchain-experimental>=0.0.20
ollama==0.1.5